# src/swagger/swagger-latest.yaml
swagger: '2.0'
info:
  contact: {}
paths:
...
  /api/v2/posters:
    post:
      description: create poster
      consumes:
        - application/json
      tags:
        - posters/v2
      summary: Create
      parameters:
        - description: Poster body
          name: input
          in: body
          required: true
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.PosterRequest'
        - type: string
          description: JWT-format token
          name: X-User-Token
          in: header
          required: true
      responses:
        '201':
          description: id
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.IDResponse'
        '400':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '500':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
  /api/v2/posters/{id}:
    get:
      description: get poster
      consumes:
        - application/json
      tags:
        - posters/v2
      summary: Get
      parameters:
        - type: string
          description: JWT-format token
          name: X-User-Token
          in: header
          required: true
        - type: integer
          description: PosterId
          name: id
          in: path
          required: true
      responses:
        '200':
          description: Poster
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.PosterResponse'
        '400':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '404':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '500':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
    put:
      description: update poster
      consumes:
        - application/json
      tags:
        - posters/v2
      summary: Update
      parameters:
        - description: Poster body
          name: input
          in: body
          required: true
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.PosterRequest'
        - type: integer
          description: PosterId
          name: id
          in: path
          required: true
        - type: string
          description: JWT-format token
          name: X-User-Token
          in: header
          required: true
      responses:
        '200':
          description: OK
        '400':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '404':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '500':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
    delete:
      description: delete poster
      consumes:
        - application/json
      tags:
        - posters/v2
      summary: Delete
      parameters:
        - type: string
          description: JWT-format token
          name: X-User-Token
          in: header
          required: true
        - type: integer
          description: PosterId
          name: id
          in: path
          required: true
      responses:
        '200':
          description: OK
        '400':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '404':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '500':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
...
definitions:
  git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse:
    type: object
    properties:
      error:
        type: string
        example: error description
  git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.IDResponse:
    type: object
    properties:
      id:
        type: integer
...
  git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.PosterRequest:
    type: object
    properties:
      chrono:
        type: integer
      genres:
        type: array
        items:
          type: string
      name:
        type: string
      year:
        type: integer
  git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.PosterResponse:
    type: object
    properties:
      chrono:
        type: integer
      createdat:
        description: will not be used, satisfy musttag linter
        type: string
      genres:
        type: array
        items:
          type: string
      id:
        type: integer
      name:
        type: string
      userId:
        type: integer
      year:
        type: integer
...